Function renaming. Removal of scalem_2d file.
