+++ src/mmg3d/mmg3d.h
@@ -580,6 +580,7 @@ void MMG3D_Set_commonFunc(void) {
   MMG5_indElt            = MMG3D_indElt;
   MMG5_grad2met_ani      = MMG5_grad2metSurf;
   MMG5_grad2metreq_ani   = MMG5_grad2metSurfreq;
+  MMG5_solTruncature_ani = MMG5_3dSolTruncature_ani;
 
 #ifdef USE_SCOTCH
   MMG5_renumbering       = MMG5_mmg3dRenumbering;
